+++ BackEnd/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Boolean, DateTime, JSON, func
+from sqlalchemy import Column, Integer, String, Float, Boolean, DateTime, JSON, func, Numeric
 from sqlalchemy.orm import declarative_mixin, declared_attr
 
 from db import Base
@@ -10,9 +10,9 @@ class BaseDeal:
     section = Column(Integer, nullable=False)
     stage = Column(Integer, nullable=False)
     address = Column(String, nullable=False)
-    sqft = Column(Float, nullable=True)
+    sqft = Column(Numeric(10, 2), nullable=True)
     bedrooms = Column(Integer, nullable=True)
-    bathrooms = Column(Float, nullable=True)
+    bathrooms = Column(Numeric(4, 2), nullable=True)
     zillow_link = Column(String, nullable=True)
     overall_design = Column(String, nullable=True)
     crime_rate = Column(String, nullable=True)
@@ -35,18 +35,18 @@ class BaseDeal:
     
     notes = Column(String, nullable=True)
 
-    purchase_price_in_thousands = Column(Float, nullable=False)
-    rehab_cost_in_thousands = Column(Float, nullable=False, default=0.0)
-    down_payment = Column(Float, nullable=False)
-    closing_costs_buy_in_thousands = Column(Float, nullable=False, default=0.0)
+    purchase_price_in_thousands = Column(Numeric(12, 2), nullable=False)
+    rehab_cost_in_thousands = Column(Numeric(12, 2), nullable=False, default=0.0)
+    down_payment = Column(Numeric(5, 2), nullable=False)
+    closing_costs_buy_in_thousands = Column(Numeric(12, 2), nullable=False, default=0.0)
     use_HM_for_rehab = Column(Boolean, nullable=False, default=False)
-    HML_points = Column(Float, nullable=False, default=0.0)
-    HML_interest_rate = Column(Float, nullable=False)
+    HML_points = Column(Numeric(5, 2), nullable=False, default=0.0)
+    HML_interest_rate = Column(Numeric(5, 2), nullable=False)
     
     # Holding costs (Shared mostly)
-    annual_property_taxes = Column(Float, nullable=False, default=0.0)
-    annual_insurance = Column(Float, nullable=False, default=0.0)
-    montly_hoa = Column(Float, nullable=False, default=0.0)
+    annual_property_taxes = Column(Numeric(12, 2), nullable=False, default=0.0)
+    annual_insurance = Column(Numeric(12, 2), nullable=False, default=0.0)
+    montly_hoa = Column(Numeric(12, 2), nullable=False, default=0.0)
     
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now())
@@ -56,30 +56,30 @@ class BrrrActiveDeal(Base, BaseDeal):
     __tablename__ = "active_deals"
 
     # BRRRR Specific
-    arv_in_thousands = Column(Float, nullable=False) # Used for Refi LTV
-    Months_until_refi = Column(Float, nullable=False)
-    closing_cost_refi_in_thousands = Column(Float, nullable=False, default=0.0)
+    arv_in_thousands = Column(Numeric(12, 2), nullable=False) # Used for Refi LTV
+    Months_until_refi = Column(Numeric(5, 1), nullable=False) # Can be half month?
+    closing_cost_refi_in_thousands = Column(Numeric(12, 2), nullable=False, default=0.0)
     loan_term_years = Column(Integer, nullable=False, default=30)
-    ltv_as_precent = Column(Float, nullable=False)
-    interest_rate = Column(Float, nullable=False) # Long term loan rate
-    rent = Column(Float, nullable=False)
-    vacancy_percent = Column(Float, nullable=False, default=0.0)
-    property_managment_fee_precentages_from_rent = Column(Float, nullable=False, default=0.0)
-    maintenance_percent = Column(Float, nullable=False, default=0.0)
-    capex_percent_of_rent = Column(Float, nullable=False, default=0.0)
+    ltv_as_precent = Column(Numeric(5, 2), nullable=False)
+    interest_rate = Column(Numeric(5, 2), nullable=False) # Long term loan rate
+    rent = Column(Numeric(12, 2), nullable=False)
+    vacancy_percent = Column(Numeric(5, 2), nullable=False, default=0.0)
+    property_managment_fee_precentages_from_rent = Column(Numeric(5, 2), nullable=False, default=0.0)
+    maintenance_percent = Column(Numeric(5, 2), nullable=False, default=0.0)
+    capex_percent_of_rent = Column(Numeric(5, 2), nullable=False, default=0.0)
 
 
 class FlipActiveDeal(Base, BaseDeal):
     __tablename__ = "flip_deals"
     
-    sale_price_in_thousands = Column(Float, nullable=False) # ARV
+    sale_price_in_thousands = Column(Numeric(12, 2), nullable=False) # ARV
     holding_time_months = Column(Integer, nullable=False)
     
-    buyer_agent_selling_fee = Column(Float, nullable=False, default=0.0)
-    seller_agent_selling_fee = Column(Float, nullable=False, default=0.0)
-    selling_closing_costs_in_thousands = Column(Float, nullable=False, default=0.0)
+    buyer_agent_selling_fee = Column(Numeric(5, 2), nullable=False, default=0.0)
+    seller_agent_selling_fee = Column(Numeric(5, 2), nullable=False, default=0.0)
+    selling_closing_costs_in_thousands = Column(Numeric(12, 2), nullable=False, default=0.0)
     
-    monthly_utilities = Column(Float, nullable=False, default=0.0)
-    capital_gains_tax_rate = Column(Float, nullable=False, default=0.0)
+    monthly_utilities = Column(Numeric(12, 2), nullable=False, default=0.0)
+    capital_gains_tax_rate = Column(Numeric(5, 2), nullable=False, default=0.0)
     
     sale_comps = Column(JSON, nullable=True)
